requirements.txt
Flask==2.3.3
requests==2.31.0
pdf2image==1.16.3
Pillow>=10.0.0
python-docx==0.8.11
Werkzeug==2.3.7
python-dotenv==1.0.0
PyPDF2==3.0.1
openai==1.3.0